images addet and other stuff

+++ app/src/main/java/model/StatProperty.java
@@ -15,7 +15,8 @@ public class StatProperty {
 
     public int rank;
 
-    public int percentile;
+
+    public double percentile;
 
     public String displayValue;
 }
